Add comments. Code refactoring.
