+++ AirQualityMonitoring-SwiftUI-Combine/Model/CityData.swift
@@ -9,25 +9,31 @@ import Foundation
 import Combine
 import SwiftUI
 
-struct CityDataResponse: Codable {
-    let city: String
-    let aqi: Double
-}
-
+/// A model to show this in List
+///
+/// Confirms to `Identifiable` protocol for the SwiftUI List compatibility
 class CityData: Identifiable {
     let id: UUID = UUID()
     var name: String = ""
     var aqi: Double = 0.0
 }
 
+/// `CityData` extension
+///
+/// Retrieve the other information based on air quality index value
 extension CityData {
+    
+    /// A variable - to get `AirQualityIndexClassification`
+    /// based on air quality index value
     var airQuality: AirQualityIndexClassification {
         get {
             return AirQualityIndexClassifier
                 .classifyAirQualityIndex(aqi: self.aqi)
         }
     }
-    
+        
+    /// A variable - to get `Color`
+    /// based on Air Quality Classification
     var aqiColor: Color {
         get {
             return AirQualityIndexInfoClassifier
@@ -35,7 +41,8 @@ extension CityData {
                 .suColor
         }
     }
-    
+        
+    /// A variable - to get air quality classification description text in `String`
     var aqiDescriptionText: String {
         get {
             return AirQualityIndexInfoClassifier
