+++ AirQualityMonitoring-SwiftUI-Combine/Utils/Constants.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+/// Server URL String information
 enum ServerConnection {
     public static var url: String = "ws://city-ws.herokuapp.com"
 }
