+++ AirQualityMonitoring-SwiftUI-Combine/View/CitiesListView.swift
@@ -7,12 +7,20 @@
 
 import SwiftUI
 
+/// A `View` to render the list of Cities with their Air Quality Index Information
 struct CitiesListView: View {
-    
+        
+    /// A `@StateObject` variable of `CitiesListViewModel`, which is bind to the List
+    ///
+    /// The list is updated when Air quality index is changed for any city
     @StateObject var modelData = CitiesListViewModel(with: DataProvider())
-    
+        
+    /// A default View renders with body variable of `View`
     var body: some View {
+        
+        /// A navigation-view
         NavigationView {
+            
             List {
                 ForEach($modelData.cities) { city in
                     ZStack {
@@ -36,6 +44,9 @@ struct CitiesListView: View {
     }
 }
 
+/// A `PreviewProvider` for `CitiesListView`
+///
+/// We can check that the View is rendered as expected
 struct CitiesListView_Previews: PreviewProvider {
     static var previews: some View {
         CitiesListView(modelData: CitiesListViewModel(with: DummyDataProvider()))
